Better command line help. Code cleanup.

+++ smiles-vhts/util/generate_features.py
@@ -1,6 +1,6 @@
+import math
 import h5py
 from progressbar import ProgressBar
-import math
 from keras import models
 
 
@@ -12,7 +12,7 @@ def generate_features(smiles_file, model_file, features_file, batch_size):
     model = models.load_model(model_file)
     features = features_hdf5.create_dataset('features', (smiles_matrix.shape[0], model.output_shape[1]))
     with ProgressBar(max_value=len(smiles_matrix)) as progress:
-        for i in range(math.ceil(smiles_matrix.shape[0]/batch_size)):
+        for i in range(int(math.ceil(smiles_matrix.shape[0]/batch_size))):
             start = i * batch_size
             end = min(smiles_matrix.shape[0], (i + 1) * batch_size - 1)
             results = model.predict(smiles_matrix[start:end])
