+++ smiles-vhts/util/oversample.py
@@ -1,7 +1,7 @@
 import os
+import math
 import h5py
 from progressbar import ProgressBar
-import math
 
 
 def oversample(data_file):
@@ -28,9 +28,9 @@ def oversample(data_file):
     left_difference = difference
     with ProgressBar(max_value=oversampled_classes.shape[0]) as progress:
         if class_zero_count < class_one_count:
-            copies_per_instance = math.ceil(class_one_count / class_zero_count)
+            copies_per_instance = int(math.ceil(class_one_count / class_zero_count))
         else:
-            copies_per_instance = math.ceil(class_zero_count / class_one_count)
+            copies_per_instance = int(math.ceil(class_zero_count / class_one_count))
         target_i = 0
         for i in range(len(classes)):
             minority = (class_zero_count < class_one_count and classes[i][0] >= classes[i][1]) or \
